requirements.txt
#Ce sont les seules dépendances nécessaires pour lancer l'interface et tester les pages web.

## 📦 `requirements.txt`

flask==3.0.3
requests==2.31.0
google-genai
python-dotenv==1.0.1
